// lib/types/github/github-user-data.type.ts
export type GitHubUserRepo = {
  // The entire data of the API response is wayyy larger than that.
  // This is only part of that data that I want to display on the site.jjhgjhgh
  name: string;
  description: string;
  html_url: string;
  created_at: string;
  updated_at: string;
  pushed_at: string;
  fork: boolean;
  homepage: string;
};